forwarding-powerpc-elf.S: Don't use %r0 as tmp reg
While it works with lwz, it will not work with lis and addi. The PIC
version could still use %r0, but it's easier to read when keeping it
consistent and is the same number of instructions.

FossilOrigin-Name: bacca956e5acbd84fe0cba1c96929d1306d2a796745e758bd67c1cca3e92c864