Better fallback asprintf() implementation
Instead of trying if it's C99 compliant and falling back to a 64 KB
buffer if it is not, resize the buffer in powers of 2 until it fits.

The problem is that some implementations that are not compliant with C99
don't just return -1 for a length of 0, but actually crash (e.g. HP-UX).

FossilOrigin-Name: 7267f55840e0f00423d782e94e830949f946820e5ea57841791a231165139bdf