OFZIPArchive: Make returned streams retain archive
In order to not create a retain cycle, this changes the reference to the
last returned stream to an unsafe unretained one that the stream itself
needs to reset to nil in its dealloc.

Since when closing a stream for writing the archive needs to be updated,
the code for that was moved out of OFZIPArchive and into the stream
writer, since it now has a reference to the archive anyway.

FossilOrigin-Name: 6527c97d03ed80fc1cffc2e6a98422292874db830c757023cee3963aa789797e