OFGameController: Fix Xbox 360 controller mapping
It turns out the quirk for all other controllers to swap north/west was
wrong, and the Xbox 360 controller is actually the one that has
north/west swapped.

The reason for this seems to be that Linux has BTN_NORTH and BTN_X
mapped to the same code, as well as BTN_WEST to BTN_Y. However, on the
Xbox 360 controller, Y is north and X is west. It seems that the Xbox
360 driver decided to return BTN_X and BTN_Y (so that the labels are
correct) while other drivers return BTN_NORTH and BTN_SOUTH (so that the
position is correct).

FossilOrigin-Name: a4fa1b26f20c82012ded39fcc27151390d423fda830f2327d1888a8afdb907fe