OFHTTP{Client,Server}: Don't set non-blocking
The user decides whether they want to use it blocking or non-blocking.

If it is set to non-blocking but the user is trying to perform
synchronous operations, it fails. If it is set to blocking but the user
performs asynchronous operations, it works. So blocking is the better
default.

FossilOrigin-Name: e042f71337b1b34f4bc202675c82ffd6dc77022a249faf8182a9124716edd547