Revert to generating ObjC for MorphOS library glue
The initial assembly code forgot that r13 needs to be saved. While it
was changed later to save r13, that code was still wrong, as arguments
might be passed on the stack. To properly copy arguments passed on the
stack into the new stack frame, awareness of the arguments and their
types is required. It's much easier to just generate ObjC code and let
the compiler handle it.

FossilOrigin-Name: 0d10241888a4dcb65bbea82f1d58cb2ae18f25364bfb0f5b7f7c972ef38cc43f