Disallow different read/write objects with poll
This was only half of a solution and does not solve the real problem.
The proper solution was to cancel all async requests before the object
with the same FD for the async I/O changes.

FossilOrigin-Name: bd9075284024df58458f1e65705f258d39ecd9665f40fcbc18899db32af70f4d