Use @compatibility_alias also with modules
C++ 20 and up use modules by default in newer Clang versions, which made
the compatibility aliases disappear due to the !__has_feature(modules)
and hence object literals didn't work.

In the past, compatibility aliases with modules caused problems on
macOS, but that seems to be no longer the case, so it seems to be safe
to always use them now.