OFLHAArchiveEntry: Read header 0/1 Amiga comment
Amiga LHA seems to have used a hack where it stores the Amiga comment in
the file name after a \0 while the file name length includes the
comment. This is surprising, as the extended area would have made more
sense for this and would have been safer. Luckily, -[OFStream
readStringWithLength:] always consumed the specified number of bytes
from the stream while truncating the string at a \0, so this is why
Amiga comments caused no issues before this change.