Rename OFInflateStream -> OFDeflateStream
Initially, the idea was to split compression and decompression into
OFInflateStream and OFDeflateStream to avoid always allocating instance
variables for compression and decompression. However, this is rather
unexpected for a user, so it's better to have it all in one class, but
use a pointer to a struct that is only allocated depending on the mode
used.